+++ test/bigtest/tests/unit/epics/get-kb-credentials-key-test.js
@@ -1,0 +1,90 @@
+/* global describe, it, beforeEach */
+import { expect } from 'chai';
+
+import { TestScheduler } from 'rxjs/Rx';
+
+import { createGetKbCredentialsEpic } from '../../../../../src/redux/epics';
+import {
+  GET_KB_CREDENTIALS_KEY,
+  GET_KB_CREDENTIALS_KEY_SUCCESS,
+  GET_KB_CREDENTIALS_KEY_FAILURE,
+} from '../../../../../src/redux/actions';
+
+describe('(epic) getKBCredentialsKey', () => {
+  let testScheduler;
+
+  beforeEach(() => {
+    testScheduler = new TestScheduler((actual, expected) => {
+      expect(actual).deep.equal(expected);
+    });
+  });
+
+  const state$ = {
+    getState: () => {
+      return {
+        okapi: {
+          url: 'https://folio-snapshot',
+          tenant: 'diku',
+          token: 'token',
+        }
+      };
+    }
+  };
+
+
+  it('should handle successful data fetching', () => {
+    const response = {
+      data: {
+        attributes: {
+          apiKey: 'some-key',
+        },
+      },
+    };
+
+    const action$ = testScheduler.createHotObservable('-a', {
+      a: { type: GET_KB_CREDENTIALS_KEY }
+    });
+
+    const dependencies = {
+      knowledgeBaseApi: {
+        getCredentialsKey: () => testScheduler.createColdObservable('--a', {
+          a: response.data
+        })
+      }
+    };
+
+    const output$ = createGetKbCredentialsEpic(dependencies)(action$, state$);
+
+    testScheduler.expectObservable(output$).toBe('---a', {
+      a: {
+        type: GET_KB_CREDENTIALS_KEY_SUCCESS,
+        payload: response.data,
+      }
+    });
+
+    testScheduler.flush();
+  });
+
+  it('should handle errors', () => {
+    const action$ = testScheduler.createHotObservable('-a', {
+      a: { type: GET_KB_CREDENTIALS_KEY }
+    });
+
+    const dependencies = {
+      knowledgeBaseApi: {
+        getCredentialsKey: () => testScheduler.createColdObservable('--#', null, 'Error messages')
+      }
+    };
+
+    const output$ = createGetKbCredentialsEpic(dependencies)(action$, state$);
+
+    testScheduler.expectObservable(output$).toBe('---a', {
+      a: {
+        type: GET_KB_CREDENTIALS_KEY_FAILURE,
+        payload: 'Error messages',
+      }
+    });
+
+    testScheduler.flush();
+  });
+});